+++ contracts/transfer/src/lib.rs
@@ -15,7 +15,7 @@ mod transfer;
 mod wasm;
 
 pub use error::Error;
-pub use transfer::{Call, TransferContract};
+pub use transfer::{Call, Leaf, TransferContract};
 pub type Map<K, V> = dusk_hamt::Hamt<K, V, ()>;
 
 #[cfg(target_arch = "wasm32")]